Move all changes in `numpy-openblas` and revert `numpy`
