+++ recipes/recipes_emscripten/numpy-openblas/emscripten.meson.cross
@@ -1,6 +1,6 @@
 [binaries]
 exe_wrapper = 'node'
-pkgconfig = 'pkg-config'
+pkg-config = 'pkg-config'
 
 [properties]
 needs_exe_wrapper = true
