+++ recipes/recipes_emscripten/numpy-openblas/recipe.yaml
@@ -10,8 +10,10 @@ package:
 source:
   url: https://github.com/numpy/numpy/releases/download/v${{ version }}/numpy-${{ version }}.tar.gz
   sha256: a48a113e6afea91f5608793bafa7ef2ad481fefbda87ec5069f483de61cb9fa3
+  patches:
+  - patches/0001-Match-flang-OpenBLAS-Fortran-ABI-on-Emscripten-wasm32.patch
 build:
-  number: 0
+  number: 1
 
   files:
     exclude:
@@ -35,6 +37,10 @@ requirements:
   - pkg-config
   - python
   host:
+  - openblas
+  - python
+  run:
+  - openblas
   - python
   run_exports:
   - numpy >=${{ default_abi_level }},<3
@@ -58,3 +64,4 @@ about:
 extra:
   recipe-maintainers:
   - DerThorsten
+  channel: emscripten-forge-4x-experimental
